requirements.txt
fastapi~=0.110.0
uvicorn==0.27.1
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.3.2
starlette==0.36.3
tortoise-orm==0.20.0
asyncpg
bson
pyjwt==2.10.1
passlib==1.7.4
email-validator==2.2.0
python-multipart==0.0.20
gmqtt==0.7.0
